refactor(Calculator at el): Refactor calculate function to Calculator object
