+++ src/main/kotlin/step2/Main.kt
@@ -1,7 +1,7 @@
 package step2
 
+import step2.calculator.Calculator.calculate
 import step2.calculator.Input
-import step2.calculator.calculate
 
 fun main() {
     println("사칙 연산을 수행할 문자열을 입력해주세요")
